+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Truncate app title and description in app picker
 - do not open self chat on info message click #5361
+- fix Connectivity colors in dark mode #5397
 
 <a id="2_10_0"></a>